Fix broken version JSON view
A 2017 commit made Result.type a Type rather than a string that can
immediately be used and compared: the actual type string is stored
in Result.type.type.

This should fix the version JSON view.

This also fixes a potential issue in get_version_results where
Type objects were being used as keys in a dict.